backend/requirements.txt
cognee[docs]
fastapi
uvicorn[standard]
python-multipart
python-dotenv
aiofiles
